refactor(trie): Move `TrieUpdates` from db integration crate to the generic one
Replace database specific types with an abstraction
