+++ crates/trie/trie-db/src/lib.rs
@@ -33,6 +33,3 @@ pub mod proof;
 /// The implementation of the Merkle Patricia Trie.
 mod trie;
 pub use trie::{storage_root, state_root};
-
-/// Buffer for trie updates.
-pub mod updates;
