+++ crates/trie/trie/src/prefix_set/mod.rs
@@ -4,9 +4,6 @@ use std::{
     sync::Arc,
 };
 
-mod loader;
-pub use loader::PrefixSetLoader;
-
 /// Collection of trie prefix sets.
 #[derive(Default, Debug)]
 pub struct TriePrefixSets {
