+++ crates/trie/trie/src/trie_cursor/mod.rs
@@ -26,6 +26,24 @@ pub trait TrieCursorFactory {
     ) -> Result<Box<dyn TrieCursor<Err=Self::Err> + '_>, Self::Err>;
 }
 
+/// Factory for creating trie cursors.
+pub trait TrieCursorRwFactory {
+    /// The associated error that can occur during the creation of the cursor.
+    type Err;
+    type StorageKey;
+    type StorageValue;
+    type AccountsKey;
+    type AccountsValue;
+
+    /// Create an account trie cursor.
+    fn account_trie_cursor_rw(&self) -> Result<Box<dyn TrieCursorRw<Self::AccountsKey, Self::AccountsValue, Err=Self::Err> + '_>, Self::Err>;
+
+    /// Create a storage tries cursor.
+    fn storage_tries_cursor_rw(
+        &self,
+    ) -> Result<Box<dyn TrieCursorRw<Self::StorageKey, Self::StorageValue, Err=Self::Err> + '_>, Self::Err>;
+}
+
 /// A cursor for navigating a trie that works with both Tables and DupSort tables.
 #[auto_impl::auto_impl(&mut, Box)]
 pub trait TrieCursor: Send + Sync {
@@ -44,4 +62,27 @@ pub trait TrieCursor: Send + Sync {
 
     /// Get the current entry.
     fn current(&mut self) -> Result<Option<TrieKey>, Self::Err>;
+}
+
+/// A cursor for writing into a trie that works with both Tables and DupSort tables.
+#[auto_impl::auto_impl(&mut, Box)]
+pub trait TrieCursorWrite<K, V>: Send + Sync {
+    /// The associated error that can occur on the cursor operations.
+    type Err;
+
+    /// Deletes value at the current cursor position.
+    fn delete_current(&mut self) -> Result<(), Self::Err>;
+
+    /// Deletes all values associated with key at the current cursor position.
+    fn delete_current_duplicates(&mut self) -> Result<(), Self::Err>;
+
+    /// Updates `value` currently associated with `key` if it exists.
+    /// Otherwise, inserts a new `key`-`value` pair.
+    fn upsert(&mut self, key: K, value: V) -> Result<(), Self::Err>;
+}
+
+#[auto_impl::auto_impl(&mut, Box)]
+pub trait TrieCursorRw<K, V>: TrieCursor + TrieCursorWrite<K, V, Err=<Self as TrieCursor>::Err> + Send + Sync {
+    /// The associated error that can occur on the cursor operations.
+    type Err = <Self as TrieCursor>::Err;
 }
